Rename EmbedFactory to embed
